custom_nodes/Cosmos/requirements.txt
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# Please keep requirements sorted alphabetically
av
better-profanity
einops==0.7.0
einx==0.1.3
huggingface-hub>=0.26.2
hydra-core
imageio[ffmpeg]
iopath
loguru
mediapy
nltk
peft
pillow
pytorch_retinaface @ git+https://github.com/NVlabs/Pytorch_Retinaface.git@b843f45
sentencepiece
termcolor
transformers==4.45.0
